pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-audit"
version = "1.0.0"
description = "Outil professionnel d'audit de sécurité Linux en lecture seule"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Smail Haroun"},
]
keywords = ["security", "audit", "linux", "ssh", "devsecops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "paramiko>=3.4.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "rich>=13.7.0",
]

[project.optional-dependencies]
pdf = ["weasyprint>=62.0"]
dev = [
    "pytest>=8.1.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
    "types-paramiko>=3.4.0",
    "types-PyYAML>=6.0.12",
]
all = ["sentinel-audit[pdf,dev]"]

[project.scripts]
sentinel-audit = "sentinel_audit.cli:main"

[project.urls]
Homepage = "https://github.com/Smail-Har/sentinel_audit"
Repository = "https://github.com/Smail-Har/sentinel_audit"

[tool.setuptools.packages.find]
include = ["sentinel_audit*"]

[tool.setuptools.package-data]
sentinel_audit = ["config/*.yaml", "templates/*.jinja2"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T20", "RUF"]
ignore = ["S603", "S607", "S108", "S602", "A002"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "S107", "S104"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
